Moved execute command out of utils, as it doesn't seem like the best fit for it. Also update do, dies and ressurect to not be default exports for sake of consistency
