+++ src/utils.ts
@@ -1,28 +1,7 @@
 import { Message } from 'discord.js'
-import TimeAgo from 'javascript-time-ago'
-import en from 'javascript-time-ago/locale/en'
-import { attack } from './commands/attack'
-import { defend } from './commands/defend'
-import die from './commands/dies'
-import _do from './commands/do'
-import { generate } from './commands/generate'
-import { getConfig } from './commands/getConfig'
-import { ping } from './commands/ping'
-import resurrect from './commands/resurrect'
-import { setConfig } from './commands/setConfig'
-import { snack } from './commands/snack'
-import { speak } from './commands/speak'
-import { xpBudget } from './commands/xpBudget'
 import config from './config.json'
-import { logger } from './logger'
 import Guild, { GuildSettings } from './models/guild'
 import { ValuedDescriptor } from './models/valuedDescriptor'
-
-// set up TimeAgo
-TimeAgo.addLocale(en)
-
-// Create relative date/time formatter.
-const timeAgo = new TimeAgo('en-GB')
 
 /**
  * @description Parse the message into a command and a list of arguments which have been provided
@@ -31,7 +10,7 @@ const timeAgo = new TimeAgo('en-GB')
  * args = ["Is", "this", "the", "real", "life?"]
  * @param message - this is the Discord message
  */
-function parseMessage(message: Message): [string, string[]] {
+export function parseMessage(message: Message): [string, string[]] {
     const args = message.content
         .slice(config.prefix.length)
         .trim()
@@ -42,62 +21,9 @@ function parseMessage(message: Message): [string, string[]] {
 }
 
 /**
- * @description given a message - determine what command is being run and execute it
- * @param message
+ * Turn the settings object into a more readable string to be printed in discord
+ * @param settings the settings to be printed
  */
-export async function executeCommand(message: Message) {
-    const [command, args] = parseMessage(message)
-    logger.debug(`Command parsed: '${command}'`)
-    // Check if Purrito has run out of lives. Only allow resurrect if so.
-    const guild = await findOrMakeGuild(message.guild!.id)
-    if (guild.purritoState.lives <= 0 && command !== 'resurrect') {
-        return await message.channel.send(
-            `Purrito died ${timeAgo.format(
-                guild.purritoState.timeOfDeath
-            )}. It's time to move on.`
-        )
-    }
-    switch (command) {
-        case 'attack':
-            attack(message)
-            break
-        case 'ping':
-            ping(message)
-            break
-        case 'snack':
-        case 'snacc':
-            snack(message)
-            break
-        case 'defend':
-            defend(message)
-            break
-        case 'speak':
-            speak(message)
-            break
-        case 'do':
-            _do(message, args)
-            break
-        case 'setconfig':
-            setConfig(message, args)
-            break
-        case 'getconfig':
-            getConfig(message, args)
-            break
-        case 'die':
-            die(message)
-            break
-        case 'resurrect':
-            resurrect(message)
-            break
-        case 'generate':
-            generate(message, args)
-            break
-        case 'xpbudget':
-            xpBudget(message, args)
-            break
-    }
-}
-
 export function allConfigAsMessageString(settings: GuildSettings) {
     let response = ''
     Object.keys(config.defaultSettings).forEach(key => {
@@ -109,6 +35,11 @@ export function allConfigAsMessageString(settings: GuildSettings) {
 
     return response
 }
+
+/**
+ * Find the config for a given Guild, if the guild does not exist create a new one in MongoDB
+ * @param guildId the unique ID of the Guild/Server
+ */
 export async function findOrMakeGuild(guildId: string) {
     // Try to get existing guild config
     let guild = await Guild.findOne({ guildId: guildId })
@@ -123,10 +54,19 @@ export async function findOrMakeGuild(guildId: string) {
     return guild
 }
 
+/**
+ * When given an array, pick a random element from it and return to the user
+ * @param array the array 
+ */
 export function getRandomValueFromArray<T>(array: T[]): T {
     return array[Math.floor(Math.random() * array.length)]
 }
 
+/**
+ * When generated an item, some descriptors are weighted, to give them more / less 
+ * chance to appear than others. This function will created that weighted list
+ * @param descriptors descriptors from which to created a weighted list
+ */
 export function createdWeightedList(
     descriptors: ValuedDescriptor[]
 ): ValuedDescriptor[] {
